optimize DeclarativeEnvironmentRecord allow one call for create and initialize for binding, move funciton environment related logic to.. FunctionEnvironmentRecord
