+++ Jint/Runtime/Interpreter/Statements/JintTryStatement.cs
@@ -39,8 +39,8 @@ namespace Jint.Runtime.Interpreter.Statements
                     var c = b.Value;
                     var oldEnv = _engine.ExecutionContext.LexicalEnvironment;
                     var catchEnv = LexicalEnvironment.NewDeclarativeEnvironment(_engine, oldEnv);
-                    catchEnv._record.CreateMutableBinding(_catchParamName, canBeDeleted: false);
-                    catchEnv._record.InitializeBinding(_catchParamName, c);
+                    var catchEnvRecord = (DeclarativeEnvironmentRecord) catchEnv._record;
+                    catchEnvRecord.CreateMutableBindingAndInitialize(_catchParamName, canBeDeleted: false, c);
 
                     _engine.UpdateLexicalEnvironment(catchEnv);
                     b = _catch.Execute();
